Move procspy out of vendor into probe/endpoint.
