+++ probe/endpoint/procspy/fixture.go
