+++ probe/endpoint/procspy/lsof.go
