+++ probe/endpoint/procspy/lsof_internal_test.go
@@ -12,8 +12,8 @@ func TestLSOFParsing(t *testing.T) {
 		"p25196\n" +
 			"ccello-app\n" +
 			"n127.0.0.1:48094->127.0.0.1:4039\n" +
-			"n*:4040\n": map[string]Proc{
-			"127.0.0.1:48094": Proc{
+			"n*:4040\n": {
+			"127.0.0.1:48094": {
 				PID:  25196,
 				Name: "cello-app",
 			},
@@ -23,7 +23,7 @@ func TestLSOFParsing(t *testing.T) {
 		"cdhclient\n" +
 			"n*:68\n" +
 			"n*:38282\n" +
-			"n*:40625\n": map[string]Proc{},
+			"n*:40625\n": {},
 
 		// A bunch
 		"p13100\n" +
@@ -39,28 +39,28 @@ func TestLSOFParsing(t *testing.T) {
 			"n192.168.2.111:56385->74.201.105.31:443\n" +
 			"p21356\n" +
 			"cssh\n" +
-			"n192.168.2.111:33963->192.168.2.71:22\n": map[string]Proc{
-			"[::1]:6600": Proc{
+			"n192.168.2.111:33963->192.168.2.71:22\n": {
+			"[::1]:6600": {
 				PID:  13100,
 				Name: "mpd",
 			},
-			"[2003:45:2b57:8900:1869:2947:f942:aba7]:55711": Proc{
+			"[2003:45:2b57:8900:1869:2947:f942:aba7]:55711": {
 				PID:  14612,
 				Name: "chromium",
 			},
-			"192.168.2.111:37158": Proc{
+			"192.168.2.111:37158": {
 				PID:  14612,
 				Name: "chromium",
 			},
-			"192.168.2.111:44013": Proc{
+			"192.168.2.111:44013": {
 				PID:  14612,
 				Name: "chromium",
 			},
-			"192.168.2.111:56385": Proc{
+			"192.168.2.111:56385": {
 				PID:  14612,
 				Name: "chromium",
 			},
-			"192.168.2.111:33963": Proc{
+			"192.168.2.111:33963": {
 				PID:  21356,
 				Name: "ssh",
 			},
