+++ probe/endpoint/procspy/proc.go
