+++ probe/endpoint/procspy/procnet_internal_test.go
