+++ probe/endpoint/procspy/spy.go
